+++ src/main/java/de/jeff_media/AngelChest/CommandManager.java
@@ -20,12 +20,17 @@ public final class CommandManager {
 
     public static void registerCommand(String permission, String... aliases) {
         Main main = Main.getInstance();
+        main.debug("Registering command "+aliases[0]);
+        for(String alias : aliases) {
+            if(alias.equals(aliases[0])) continue;
+            main.debug("  Alias: "+alias);
+        }
         PluginCommand command = getCommand(aliases[0], main);
         command.setAliases(Arrays.asList(aliases));
         try {
             List<String> usage = FileUtils.readFileFromResources(pathToCommandDescriptions + aliases[0]+".usage");
             List<String> description = FileUtils.readFileFromResources(pathToCommandDescriptions + aliases[0]+".description");
-            command.setDescription(ListUtils.getStringFromList(description,System.lineSeparator()));
+            command.setDescription(ListUtils.getStringFromList(description,"\n"));
             command.setUsage(ListUtils.getStringFromList(usage,System.lineSeparator()));
             command.setPermissionMessage(main.messages.MSG_NO_PERMISSION);
             command.setPermission(permission);