+++ src/main/java/de/jeff_media/AngelChest/commands/CommandList.java
@@ -1,6 +1,6 @@
 package de.jeff_media.AngelChest.commands;
 
-import de.jeff_media.AngelChest.CommandManager;
+import de.jeff_media.AngelChest.config.Permissions;
 import de.jeff_media.AngelChest.utils.CommandUtils;
 import de.jeff_media.AngelChest.Main;
 import de.jeff_media.AngelChest.utils.Utils;
@@ -28,7 +28,7 @@ public final class CommandList implements CommandExecutor {
 
 		if(!command.getName().equalsIgnoreCase("aclist")) return false;
 		
-		if(!sender.hasPermission("angelchest.use")) {
+		if(!sender.hasPermission(Permissions.ALLOW_USE)) {
 			sender.sendMessage(main.messages.MSG_NO_PERMISSION);
 			return true;
 		}