+++ src/main/java/de/jeff_media/AngelChest/commands/CommandReload.java
@@ -1,6 +1,6 @@
 package de.jeff_media.AngelChest.commands;
 
-import de.jeff_media.AngelChest.CommandManager;
+import de.jeff_media.AngelChest.Main;
 import de.jeff_media.AngelChest.config.ConfigUtils;
 import de.jeff_media.AngelChest.config.Permissions;
 import org.bukkit.ChatColor;
@@ -15,7 +15,7 @@ public final class CommandReload implements CommandExecutor  {
     public boolean onCommand(@NotNull CommandSender commandSender, @NotNull Command command, @NotNull String s, @NotNull String[] args) {
 
         if(!commandSender.hasPermission(Permissions.ALLOW_RELOAD)) {
-            commandSender.sendMessage(main.messages.MSG_NO_PERMISSION);
+            commandSender.sendMessage(Main.getInstance().messages.MSG_NO_PERMISSION);
             return true;
         }